Crate cu29_traits

Structs

CuError

Common copper Error type.

Enums

UnifiedLogType

Defines the types of what can be logged in the unified logger.

Traits

CopperListTuple

A CopperListTuple needs to be encodable, decodable and fixed size in memory.

WriteStream

Defines a basic write, append only stream trait to be able to log or send serializable objects.

Type Aliases

CuResult