Crate cu29_traits

Common copper traits and types for robotics systems.

This crate is no_std compatible by default. Enable the “std” feature for additional functionality like implementing std::error::Error for CuError and the new_with_cause method that accepts types implementing std::error::Error.

Features

std (default): Enables standard library support

Implements std::error::Error for CuError

Adds CuError::new_with_cause() method for interop with std error types

no_std Usage

To use without the standard library:

[dependencies]
cu29-traits = { version = "0.9", default-features = false }

Structs

CuCompactString

CuError

Common copper Error type.

Enums

UnifiedLogType

Defines the types of what can be logged in the unified logger.

Constants

COMPACT_STRING_CAPACITY

Traits

CopperListTuple

A CopperListTuple needs to be encodable, decodable and fixed size in memory.

CuMsgMetadataTrait

Key metadata piece attached to every message in Copper.

ErasedCuStampedData

A generic trait to expose the generated CuStampedDataSet from the task graph.

ErasedCuStampedDataSet

Trait to get a vector of type-erased CuStampedDataSet This is used for generic serialization of the copperlists

MatchingTasks

Trait to trace back from the CopperList the origin of the messages

Metadata

Represent the minimum set of traits to be usable as Metadata in Copper.

PayloadSchemas

Trait for providing JSON schemas for CopperList payload types.

WriteStream

Defines a basic write, append only stream trait to be able to log or send serializable objects.

Functions

with_cause

Creates a CuError with a message and cause in a single call.

Type Aliases

CuResult