cu29_traits

Trait WriteStream

pub trait WriteStream<E: Encode>:
    Debug
    + Send
    + Sync {
    // Required method
    fn log(&mut self, obj: &E) -> CuResult<()>;

    // Provided methods
    fn flush(&mut self) -> CuResult<()> { ... }
    fn last_log_bytes(&self) -> Option<usize> { ... }
}

Defines a basic write, append only stream trait to be able to log or send serializable objects.

Required Methods

fn log(&mut self, obj: &E) -> CuResult<()>

Provided Methods

fn flush(&mut self) -> CuResult<()>

fn last_log_bytes(&self) -> Option<usize>

Optional byte count of the last successful log call, if the implementation can report it.